requirements.txt
transformers>=4.41.2
datasets>=2.16.0
accelerate>=0.30.0
bitsandbytes
deepspeed==0.14.4
peft>=0.11.1
jsonlines==4.0.0
evaluate
ninja
jinja2
gradio>=4.0.0
pandas>=2.0.0
scipy
einops
safetensors
sentencepiece
tiktoken
protobuf
uvicorn
pydantic
fastapi
sse-starlette
matplotlib>=3.7.0
fire
packaging
pyyaml
numpy<2.0.0
tensorboard
tqdm
latex2sympy
word2number